Update boba.py
I think that changing the three lines with the method change_to_blue so that it calls the function with the specific args and kwargs could solve the issue with the balls not turning blue when they're supposed to be. Also, the method for change_to_blue doesn't change the actual ball because it never alters the RGB like it does in the lines above it. For it to change the color of the ball to blue when the method is called, the circle would have to be changed via red=0, green=0, and blue=225 in the method.